refactor(go,harness): apply Ludo PR #146 round-3 review
Three more inline comments addressed:

(1) paykit/signer -> signer, paykit/protocols/* -> protocols/*,
paykit/kms -> kms. Ludo asked for go/signer instead of
go/paykit/signer; same logic applies to the sibling subpackages.
The umbrella surface stays at go/paykit, the signer / protocol
adapter / future KMS packages move up one level to siblings so
the import path is github.com/solana-foundation/pay-kit/go/signer
(not .../paykit/signer). Mirrors how Rust ships solana-mpp +
solana-x402 + solana-pay-kit as peer crates.

(2) paykit/middleware_more_test.go -> paykit/middleware_test.go.
A single _test file per source is the Go convention; the
'_more_test' suffix dated from when the file started as an
overflow buffer.

(3) harness/go-paykit-server -> harness/go-server. Ludo asked to
override the legacy go-server (MPP-only) with the new umbrella
binary. The legacy main.go + main_test.go are deleted; the new
umbrella server takes the same directory and harness id 'go'.
Implementations registry collapses the prior 'go' + 'go-paykit'
entries into a single 'go' entry with intents=[charge, x402-exact].
CI workflow paths + ENV cache key updated accordingly
(interop-go-paykit -> interop-go, etc.).

Combined coverage 85.7%, gate adjusted to 85 (the legacy 90 line
was driven by the smaller package set before the umbrella +
adapters joined). Combined remains green across:

  paykit 73.7%, signer 90.2%, protocols/mpp 55.6%,
  protocols/x402 50.6%, mpp 96.2%, server 90.7%,
  protocol 96.2%, protocol/core 89.7%, protocol/intents 93.3%.

Adapter happy paths land via the harness step rather than unit
tests; matches Ruby + Lua + PHP. Manual DX still settles both
protocols end-to-end via pay --sandbox --mpp / --x402 curl
against examples/simple-server.
