+++ go/examples/simple-server/main.go
@@ -16,9 +16,9 @@ import (
 	"net/http"
 
 	"github.com/solana-foundation/pay-kit/go/paykit"
-	_ "github.com/solana-foundation/pay-kit/go/paykit/protocols/mpp"
-	_ "github.com/solana-foundation/pay-kit/go/paykit/protocols/x402"
-	_ "github.com/solana-foundation/pay-kit/go/paykit/signer"
+	_ "github.com/solana-foundation/pay-kit/go/protocols/mpp"
+	_ "github.com/solana-foundation/pay-kit/go/protocols/x402"
+	_ "github.com/solana-foundation/pay-kit/go/signer"
 )
 
 func main() {
