+++ go/paykit/middleware_test.go
@@ -7,9 +7,9 @@ import (
 	"testing"
 
 	"github.com/solana-foundation/pay-kit/go/paykit"
-	_ "github.com/solana-foundation/pay-kit/go/paykit/protocols/mpp"
-	_ "github.com/solana-foundation/pay-kit/go/paykit/protocols/x402"
-	_ "github.com/solana-foundation/pay-kit/go/paykit/signer"
+	_ "github.com/solana-foundation/pay-kit/go/protocols/mpp"
+	_ "github.com/solana-foundation/pay-kit/go/protocols/x402"
+	_ "github.com/solana-foundation/pay-kit/go/signer"
 )
 
 func TestRequireFuncGateResolutionError(t *testing.T) {
