+++ go/paykit/paykit_test.go
@@ -8,9 +8,9 @@ import (
 	"testing"
 
 	"github.com/solana-foundation/pay-kit/go/paykit"
-	_ "github.com/solana-foundation/pay-kit/go/paykit/protocols/mpp"
-	_ "github.com/solana-foundation/pay-kit/go/paykit/protocols/x402"
-	"github.com/solana-foundation/pay-kit/go/paykit/signer"
+	_ "github.com/solana-foundation/pay-kit/go/protocols/mpp"
+	_ "github.com/solana-foundation/pay-kit/go/protocols/x402"
+	"github.com/solana-foundation/pay-kit/go/signer"
 )
 
 func disabled() *bool { f := false; return &f }
