+++ go/paykit/preflight_test.go
@@ -8,7 +8,7 @@ import (
 	solana "github.com/gagliardetto/solana-go"
 	"github.com/gagliardetto/solana-go/rpc"
 	"github.com/solana-foundation/pay-kit/go/paykit"
-	"github.com/solana-foundation/pay-kit/go/paykit/signer"
+	"github.com/solana-foundation/pay-kit/go/signer"
 )
 
 // fakeRPC is the paykit.PreflightRPCInterface test double. Mirrors
