+++ go/signer/signer.go
