+++ harness/go-server/go.mod
@@ -1,4 +1,4 @@
-module github.com/solana-foundation/mpp-sdk/harness/go-server
+module harness/go-server
 
 go 1.26.1
 
@@ -26,6 +26,7 @@ require (
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/mostynb/zstdpool-freelist v0.0.0-20201229113212-927304c0c3b1 // indirect
 	github.com/mr-tron/base58 v1.2.0 // indirect
+	github.com/shopspring/decimal v1.4.0 // indirect
 	github.com/streamingfast/logging v0.0.0-20250404134358-92b15d2fbd2e // indirect
 	go.mongodb.org/mongo-driver v1.17.3 // indirect
 	go.uber.org/multierr v1.11.0 // indirect
@@ -38,5 +39,3 @@ require (
 )
 
 replace github.com/solana-foundation/pay-kit/go => ../../go
-
-replace github.com/gagliardetto/solana-go => github.com/lgalabru/solana-go v0.0.0-20260403020633-3cb13b392078
